Moved RPCBased communication layer to providers
