+++ perceval/providers/kipu/kipu_communication_layer.py
@@ -26,12 +26,12 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
-from perceval.runtime.communication_layer import RPCBasedCommunicationLayer
 from perceval.serialization import InputArchive, Serialization
 
 from perceval.utils.logging import get_logger, channel
 
 from .kipu_rpc_handler import KipuRPCHandler
+from ..rpc_based_communication_layer import RPCBasedCommunicationLayer
 
 
 class KipuCommunicationLayer(RPCBasedCommunicationLayer):
